Restructured code into multiple compenents, improved readability and added documentation
